feat: compute_contract_hash
this is based on the cairo-lang v0.7.0 implementation in python and
cairo. it leans heavily on serde_json::Value for (de)serialization and
thus is quite expensive. only as little as possible of the contents of
contract definition are being understood to be able to compute the hash.

ideas attempted before:

1. using only serde_json::Value
2. learning of the full structure

first suffers from contract_definition.program.hints ordering
differently -- those are ints in python, so not ordered by string
ordering.

second suffers from the inability to find all of the pieces from the
python codebase, some top level values are `Any` which makes this
impossible, even if parts seem to have quite uniform structure.

includes tests for the used serde_json features.