fix(sync): new contracts missing from global state
This fixes a bug in the sync process. Newly deployed contracts should be
added to the global state tree with a contract root of ZERO. This was
not happening.

This went unnoticed thus far because of a Starkware bug where contracts
without any storage changes during deployment, would still write ZERO to
a field. This meant that even though we weren't init these contracts to
ZERO for deployment, the same update would contain storage updates
forcing the contract root to ZERO anyway. Once this bug was fixed, our
bug became an issue.