fix(ext_py): unconditionally trace stderr from py
this could be better, but not easily: the best possible implementation
would bind all read lines to the span we got from the channel. however,
AsyncBufReaderExt::read_line was not cancellation safe so it shouldn't be
executed on the side. it could be that sharing a tokio::sync::watch
channel for the latest span, or no span is the way to go to avoid
cancelling the read_line with tokio::select!.