feat(pathfinder): work without a Python virtualenv
`call.py` is now included in the `pathfinder` executable and we're creating
temporary script files to execute the included source with a Python
interpreter.

This way we're not using the VIRTUAL_ENV environment variable anymore
and only depend on the PATH being set up properly so that running
`python3` results in a Python interpreter set up with our `cairo`
dependencies.

This makes it possible to just install those Python dependencies globally
in a Docker image and run `pathfinder` without a virtualenv.