fix: replace zstd::bulk::Decompressor usage
in favor of zstd::decode_all. at the moment we cannot use bulk
decompressor to decode, because we would either need a safe large
allocation, leading to overestimation or having to update the estimates.
this way is slower.

this can be later fixed by including for example the uncompressed size
for each compressed payload, which would allow to use exact sized
preallocations.