feat(state): add block hash computation
Add computation of block hashes, as documented in the
[official docs](https://docs.starknet.io/docs/Blocks/header/).

The specification is slightly misleading and incorrect in some places.

Transaction commitment is not computed as a Merkle tree of the
transaction hashes, but with a Pedersen hash value that combines the
transaction hash _and_ the hash of the signature array.

Also, we know that some block ranges used a sequencer address for
block hash computation that has been lost since and the sequencer
API returns invalid seqeuncer address values for those blocks.