refactor: rewrite errorcode partialeq with convs
instead of rebuilding the errorobject in the partialeq and duplicating
code, rebuild the jsonrpsee::...::Error from the ErrorCode or hardcoded
EventFilterError.

also 1) scopes down the `use` for jsonrpsee::...::Error because the file
has too many items with "error" in their name and 2) adds a debug
assertion for using From<ErrorCode> for jsonrpsee::...::Error with the
PageSizeTooBig. perhaps there's a better idea to be found for this.