refactor(github/docker): tag manually triggered builds as snapshots
This change introduces the use of `docker/metadata-action` to determine
which tags we should tag the Docker image with.

The action provides a flexible way of defining rules to determine tags,
and we're using the following:

* For Git tags that are pushed we're using the name of the Git tag as
  the Docker tag.
* In all other cases we're tagging with `snapshot-{{branch}}-{{sha}}`.

`latest` is managed automatically by `docker/metadata-action`: builds
triggered by Git tags are automatically tagged as `latest`, too.