fix(storage): add an integer primary key to the starknet_events table
This is required so that we can stop relying on non-stable rowids in
the FTS5 virtual table we're using for indexing keys.

According to https://www.sqlite.org/rowidtable.html:

> If the rowid is not aliased by INTEGER PRIMARY KEY then it is not
> persistent and might change. In particular the VACUUM command will
> change rowids for tables that do not declare an INTEGER PRIMARY KEY.
> Therefore, applications should not normally access the rowid directly,
> but instead use an INTEGER PRIMARY KEY."

This change adds a migration that re-creates the starknet_events table
with an explicit INTEGER PRIMARY KEY and then re-creates all indexes
for the events.

Fixes issue #353.