fix: hoist spans into spawn_blocking
this will ensure the calling rpc_method=name span will exist in both the
async and sync part, should there be need for it.

checked that tokio doesn't at least yet provide any context hoisting
like this, but it can wrap the blocking task in it's own span with the
tracing feature.