fix(sequencer): make URL query string optinoal in warp mock server
Previously our only route configured on the mock server actually
required query parameters to be present.

This changes the setup so that the query string is optional now. Note
that we still "canonicalize" the full path and query and append a `?`
even if there are now query parameters. This is actually required
so that we don't have to modify tests that use these kind of URLs.