fix(py): use call_raw
call_raw is purpose-built for calls (which calls and fee est do). it
bypasses some of the things blocking us (fee charging). this required
some back and forth, and ended up with an inlined call_raw which
hopefully works. we should get back to the upstream version on next
patch release.

version zero is probably needed for now, unsure what that is for.