fix(rpc): fix return type of `starknet_getClass` and `starknet_getClassAt`
According to the OpenRPC specification the Base64 encoded, gzip
compressed JSON representation of the class should be returned. (The
same representation that is used by the write API.)

Just like the entry points, we don't store this in the database
directly, so we have to decompress our full JSON, extract the relevant
part, and then compress and encode it as expected by the specification.