feat(ext_py): hoist more spans, stderr as errors
now the span which is active at the time of calls to
Handle::{call,estimate_fee} will most likely get to be the parent span
of the stderr output, making it possible to correlate any stderr
messages.

the stderr messages are parsed as json if they start with optional
whitespace followed by json string start ", which should allow easy
multiline messages. a bit wasteful maybe.

the hoisted parent span is communicated through a Arc<Mutex<Span>> which
is not expected to be contented since most writes will happen strictly
after rust side sends the line and before the python side sends the
response.