feat: download declared classes at start
this doesn't seem to be so horribly long operation, on mainnet it's
ten-ish seconds at the moment but testnet will of course take more time.

the 3 threaded seems most efficient at least for thread utilization.
download is slowest part of course.

the progress indicator just checks the elapsed time on every row, which
is lazy solution, but works. expecting less than 10 reports for most
installations.