fix(py): lower logging levels
original idea was that calls and fee estimations are rare, we should log
them at error. after looking more into filtering with
tracing_subscriber, i learned one cannot turn off by span, only to
enable them, which is why these messages need to be under the default
level of INFO in order to be opted-in.