feat: starknet_version tracking and accepting
it was introduced to get_block responses in 0.9.1 as metadata. it is not
yet read back in any way. as we have no guarantees on the format of the
data but will very likely repeat, it was decided to be stored in a
"interning" table instead of inline on starknet_blocks.

Co-authored-by: Krzysztof Lis <chris@equilibrium.co>