fix(ext_py): avoid BrokenPipe exceptions in Python subprocesses
When cancelling a request (due to the JSON-RPC client going away before
we've finished processing) we kill() Python subprocesses, this is part
of normal operation. However, currently we close stdin and stdout, wait
one second, and only then proceed with actually sending SIGKILL to the
subprocess.

In case the Python subprocess finishes processing the command when we
have already closed stdout then sending the result of the command will
fail with a BrokenPipe exception, leading to the exception being logged
as ERROR in the pathfinder log.

By removing the explicit close we're now closing stdout _only_ after
actually killing the subprocess. This avoids the BrokenPipe exception
because in the worst case (when there's not enough buffer space for
the result in the pipe we're using as stdout) the print() will just
block in the Python process before we're killing the process -- no
exception raised.

Fixes #466