feat(py): default to latest if block is not found with pending changes
basically working around the fact that we cannot do this on the rust
side, but the best we can do is to run the python code on most accurate
block with the pending updates. however with that, reorg could come and
delete that block, leaving us to either error on NoSuchBlock or handling
this internally. since the RPC api doesn't have an error code for
"please retry" we chose to handle it internally.