chore: derive eq because clippy
i disagree with clippy on the need for opportunistic Eq derivations. i'd
rather see the compiler automatically generate them, if there's any
value. but I think this is more maintainable in the long run rather than
a bunch of #[allow(clippy::..)] which no one remembers what for were
they for.