fix: unmigrate 17
for any release code, leaving leafs is ok. the amount of rows in
goerli.sqlite is so large that a multiple gigabytes of WAL is needed.
alternatively smaller transactions of deletes of 10k rows seem to work,
after creation of a partial index on empty data.

the migration would had saved 1GB out of 82GB for goerli.