fix(storage/schema): add migration step fixing starknet_events schema
Our revision 12 step failed to do the schema upgrade in case the events
table was empty.

This change adds a new step that checks if we have the explicit `id`
primary key in the `starknet_events` table -- and re-does the actual
schema migration step of revision 12 if it has not been done yet.