fix(rpc): use from_block and to_block in event filter
v0.1.0 of the JSON-RPC specification requires `from_block` and
`to_block` in the event filter object. We somehow failed to make
this change in pathfinder and were using the old `fromBlock`
and `toBlock` names.

This change keeps the old names as aliases so that we don't break the
API for users who have updated their code to account for the bug.

Fixes #536.