fix(ext_py): set nonce to None for version 0 transactions for estimateFee
cairo-lang code checks that nonce is None for version 0 transactions when
we're doing estimateFee.

We did have a check for version 0 transactions in the ext_py code, but
we did not handle the QUERY_VERSION_BASE offset (that's used by clients
when submitting transactions for `starknet_call` and
`starknet_estimateFee`).

This change makes sure that we properly check for version 0.