feat(cairo/ext_py): change gas_price semantics
Instead of `gas_price` being an Option in the `ChildCommand` we're sending
to the Python subprocess we now use the 0 extremal value to represent
the behaviour that the gas price of the current block is to be used.

This makes de-serialization easier on the Python side, and since 0 isn't
a valid/useful price anyway this is not a big loss.