fix(rpc/v02/types): fix program serialization
The `program` attribute of the contract definition is supposed to contain
a gzipped-then-base64-encoded string of the JSON defining the program.

Previously we were converting the JSON representation to a string and
were compressing that string as a JSON string type. This caused
deserialization issues when `call.py` was trying to de-serialize the
transaction.