feat(py): handle different weights for different starknet_versions
The Starknet 0.10.2 update introduces a new set of resource fee weights
used for fee estimation.

The expected behavior is to use the correct set of weights based on
the starknet version string stored for blocks.

This change uses the version number from the result of our block
resolution to set the proper weights in the GeneralConfig.