refactor(Dockerfile): use cargo-chef for build-dependency caching
Our own hack tried to mimic what cargo-chef does: creating a fake project
that has the exact same dependencies as our main one, compiling those
dependencies and then compiling the pathfinder binary using those
pre-compiled deps as a separate step. This is to allow the dependencies
to be cached as a layer so that those are not re-compiled every time.

This change modifies the Dockerfile to just use cargo-chef that does all
this in a maintanable manner.