feat(python): use our Rust Pedersen hash implementation in Python
This commit adds a crate implementing a Python module that wraps our
Pedersen hash implementation. We have found that implementation to be
significantly faster than the one that's included in cairo-lang.

Unfortunately cairo-lang doesn't allow injecting our own hash function
in some places, so we're back to monkey-patching the implementation. If
we can load our pathfinder_starkhash module we do the monkey-patching.
Otherwise the module might not be available so we continue with the
original implementation for the hash functions.

A CI job making sure that the Python module can be compiled is also
added. We actually install the generated Python wheel file when running
the Python tests to make sure it's still compatible with the cairo-lang
code.