fix(storage): properly initialize connections from the pool
SQLite requires some pragmas and configuration options to be set on each
and every connection we're having to the database.

Previously we've only set up these flags during migrating the databse,
that is, on one single connection we might have in our connection pool.

This change fixes that by moving setup steps into a function and setting
up SqliteConnectionManager to call that as an initialization step for
each connection it's creating.