feat(storage): add FTS5 index for JSON-RPC 0.3 starknet_getEvents semantics
The new JSON-RPC 0.3 `starknet_getEvents` semantics means we need to be
able to construct FTS5 expressions that match on keys in specific
positions in the event.

To support this this change adds a new FTS5 index which contains
index-prefixed keys for events: ie. (index, key) pairs for each key
position in every event. Base32 encoding is used to construct an ASCII
representation of the 33 byte binary blob of the (index, key) pair.

An externally defined SQLite scalar function is used to convert the
Base64 representation we use for the keys in the `starknet_events` table
into a space-separated string of Base32-encoded index-prefixed keys.