feat(rpc/ext_py): send class hash hints for fee estimation
Because class hash calculation is implemented in Cairo, our Rust
implementation is a few orders of magnitude faster. Unfortunately the
Python code calculates class hashes for the declared class when we
instantiate an internal transaction object, and this calculation may
take a long time.

Instead of re-calculating class hashes in Cairo, pathfinder now sends
a list of optional class hash hints for the estimate fee transaction
list and pre-fills the the class hash cache used by the cairo-lang
Python code with those values.