fix(storage): set journal mode only once
Since the WAL journal mode is persistent we should only set it once
as the first step of opening the database.

Previously it was part of the initialization code we've been running
for each and every connection (right after our connection pool has
created the connection). Unfortunately that might lead to unintended
behavior when starting up pathfinder with a non-existent database file.

The connection pool is creating multiple idle connections in parallel
right after creating the pool. Setting `journal_mode` locks the database
so _some_ of those connection setup a attempts might get a "database is
locked" error until setting the `journal_mode` to WAL succeeds.