fix(pathfinder): remove class hash verification for Cairo 0 classes
After the recent incident it became apparent that even subtle
differences between the Python JSON encoder and serde-json can lead
to production outages because of differences in how we compute class
hashes.

This change removes the check that our computed class hash matches
the class hash we were expecting and converts the previous error
stopping the sync process to a warning.