feat(rpc/get_events): return error for event filter matching too many events
We now return an internal error ("event filter matches too many events")
for event filters that are way too broad to avoid executing SQL queries
that take forever to evaluate.

The idea is that the event filter should be specific enough either in the
block range specified or the key filter so that the total number of
events matching (not considering paging) is reasonable.