chore(github): free up some space on Github Actions test runner
With some new dependencies added for the Rust-based executor we're
running out of space on the GitHub actions runner when compiling
tests.

This new step re-configures the runner so that we have more free
disk space available by removing some unused pre-installed software
and adding in an unused temporary disk to be used for the workspace.