feat(rpc/types/class): add methods to JSON-serialize class definitions
Required so that we can convert class definitions to the blockifier
representation in the executor.

Broadcasted declare transactions contain a class definition that we have
to convert to the blockifier representation. Since this conversion is not
particularly frequent the easiest (and pretty slow) way of doing so is
by transforming the class definition back to JSON and re-parse it using
the blockifier code.