fix(storage): really set "synchronous" pragma
"synchronous" is not persistent through sessions. We've previously
set "synchronous" when setting the journal mode -- but we _do_ close
the database connection after running migrate, so "synchronous" was
reset to default when the connection pool was creating new connections.

This change moves the setting to setup_connection() which gets called
as an init function from within the connection pool -- thus makes sure
that it's correctly set for the journal mode.