fix(rpc/class): work around starknet_api Cairo 0 class parsing issue
starknet_api::deprecated_contract_class::Program requires `debug_info`
to be present in the program. That should actually be optional and now
this causes a regression where declare transactions with Cairo v0 classes
with no `debug_info` present are rejected by pathfinder.

This change adds a workaround for that issue: adds an empty debug info
to the program if it's not present in the input before parsing the class
with starknet_api.

Closes #1371