fix(sync): return state update as well when receiving a non-pending block in poll_pending
poll_pending() already had an optimization: in case it was receiving a
full block instead of pending it returned the data it had to the l2
sync code which could then process the block data without re-fetching
it from the feeder gateway.

Unfortunately the code was _not_ returning the state update even though
we're now fetching the block data _and_ the state update for the pending
block with a single call.