fix(storage/event): improve performance of getEvents queries without a key filter
If there is no key filter specified in the event filter
`select_query_stragegy()` will always return `BlockRangeFirst`.

Previously it still determined the number of events covered by the block
range query but this is unnecessary since it has no effect on the result
and is quite slow for huge ranges.