feat(executor): use blockifier main-v0.12.3 branch
NOTE: this is supposed to be a temporary measure until fixes are merged
back to the blockifier branch for 0.13.0.

This commit downgrades our blockifier dependency to use the `main-v0.12.3`
branch. This branch will later contain a fix for the query version issue
but is using a much older `starknet_api` dependency which we need to
adapt our code to.

We also have to downgrade a few third party dependencies (like `serde`
and `time`) so that they are compatible with this branch of starknet_api
and blockifier.