fix(executor): add block hash value to the state for trace execution
Historical block hashes (up to block_number - 10) should be available in
the state with contract_address=1, key=block_number. This is not a
problem for non-trace methods as there we use the state from the block
specified (which already has this update from system contracts), but for
tracing we're setting up our state reader with the state at the previous
block.

Because of this we're missing the block hash value for block_number - 10
specifically which might cause execution issues with contracts using the
`get_block_hash` system call.

Closes #1547