feat(executor/simulate): `simulate()` now returns reverted traces
`simulate()` used to return an error upon the first reverted transaction.

With the JSON-RPC 0.6.0 API we've agreed to return traces with the
reverted reason instead.

Post-execution checks will be added to `starknet_simulateTransactions`
implementations for older JSON-RPC versions that check if there were
reverted transactions and throw the same error they were mapping
`CallError::Reverted` to.