fix(storage/events): add upper bounds on queries used to select strategy
Since we already have a cost limit on the event queries there's no point
in counting the exact number of matching events by the keys or range.

It's sufficient to be able to count up to reaching a limit where the
cost would be too much anyway.

This avoids issues with queries using keys of which there are a few
hundred million events in the database.