fix(rpc): add workaround for gateway not returning deployed contract address
It looks like the Starknet gateway's `add_transaction` endpoint does not
return the address of the deployed contract in the response for
DEPLOY_ACCOUNT v3 transactions.

We don't really need that address in the response: it's computed from
data available in the transaction so we can also just compute that
ourselves.