feat(load-test): update `call` task
The `starknet_call` we're now making is calling the `balanceOf` entry
point of the `StarkGate: ETH` contract to get the balance of an actual
account.

This call is much more indicative of performance than the test contract
we've been previously using.

Also, we're also using the `pending` block as that's what actual clients
do to get the current balance.