feat(executor): add `PendingStateReader`
Instead of injecting pending state updates into the `CachedState` we
now have a separate layer that is responsible for reading the pending
state update.

This allows us to avoid having to call `get_nonce` on all contracts
that have been updated in the pending update so that we can call the
right number of `increment_nonce` calls on the `CachedState`.