feat(storage): add per-block Bloom filter for events
This change adds a new table storing Bloom filters for events emitted
in that block. The filter stores the keys (prefixed with the index) and
the address of the contract that emitted the event and can be used to
avoid loading the receipts for blocks that do not contain matching
events.