feat(storage/event): return continuation token if bumping into limits
When hitting the page scan limit the code now properly returns a
continuation token to the client (along with the events already collected).

This makes it possible for clients to eventually collect all events
they are interested in by making multiple requests -- while keeping
a reasonable limit on how long a single `starknet_getEvents` request
takes.