fix(pathfinder): make polling for new releases non-critical
The release monitoring task crashes if pathfinder is started with a
non-semver version or there's an issue setting up the HTTP client to
GitHub. This makes the task non-critical, so that it doesn't crash
pathfinder if it fails.

A log message is added to indicate that we won't poll for new releases
but otherwise the errors are ignored.

Closes #1686