fix(merkle-tree,storage): don't re-use storage indices
Previously we were using a HashMap to return new nodes when committing
the Merkle trie and hash values as unique ids for child nodes.

Unfortunately this might have lead to reusing the same
storage index multiple times if the Merkle tree had subtrees inserted
which were exactly the same.

Normally this is not a problem -- but this breaks the assumption that we
can safely delete a node from the storage if it's not referenced by its
parent anymore.

This commit changes to the code to use vectors to return new nodes.
Child nodes are referenced by their index in the vector.