fix(gateway-client): use a more aggressive timeout
Individual requests to the feeder gateway can sometimes take a long time
to complete (we've observed ~30s for polling the pending block).

In these cases it can be generally better to just time out the request
more aggressively (after just 5 seconds) and retry.

This change has to be coupled with more aggressive settings in our
retry backoff strategy.