feat(executor): enable L1 blob data availability mode for execution
If the block has its `l1_da_mode` set to blob we now enable using it
in the block context for blockifier.

Pre-v0.7 JSON-RPC estimate/simulate methods purpusefully disable this
feature even if the block we're executing on has it enabled.

This is because these methods do not return `data_gas_consumed` so
`resource_bounds` would not be initialized properly based on that
partial fee estimate.

Partially implements issue #1782.