fix(storage/event): fix handling of page overflow for last block in range
Previously we did the is-page-full check _after_ increasing the current
block number and checking if we're out of the range in the filter.

This causes the code to behave like it's done scanning the block range
even if not all events from the last block fit into the page.

This change fixes the issue by moving the check right after the
potentially page-filling `scan_block_into()` call.