chore: Fix clippy warnings when running with no `p2p` feature enabled
When running bare `cargo clippy` the tool complains about the unit arg
which is used when the `p2p` feature is disabled. As the CI runs clippy
with all features the warning is not visible there.

This patch silences clippy's `unit_arg` lint but only of the `p2p`
feature is enabled.