fix(merke-tree): don't re-create tree root node when committing
If the root node is unresolved then the Merkle tree has been created
but no update operations have been made.

In this case the previous code caused `commit()` to always re-create the
root node so that opening a tree, then committing without changes added
a new root node.

This change makes sure we return the current root commitment and node
index in that case instead.