fix(storage/schema): allow duplicate trie removals rows for all tries
Our schema has previously allowed multiple contract trie removals rows
because the contracts "trie" is in fact multiple tries updated during
separate updates.

This change now does the same for the class and storage tries as well.
The reason we have to do this is that during a reorg we're moving trie
node removal data from the reorged-away block range to the reorg target
(the last block we're keeping).

The easiest way to implement that is updating the block number in the
already existing rows -- alternative implementations would require
loading, de-serializing, merging, and re-adding the data to storage.