fix(pathfinder/reorg): re-parent deleted trie node data to the reorg target
In general, removing the trie nodes deleted at block N is only safe if we
don't ever need to access trie state at block < N. This is not necessarily
the case during a reorg: if our reorg/revert target is still in the range
of blocks we're keeping trie history for then removing deleted nodes for
the reorged-away blocks would break the trie for the blocks _before_ the
reorg target. Instead, we move all the removed nodes in the reorged-away
range to be "owned" by the revert target block (last block we're keeping).