feat(storage): use zstd dictionaries to compress transaction related data
Using properly trained zstd dictionaries seems to improve our compression
ratio by ~10% on a mainnet database. For the Sepolia testnet the
improvement is ~5 -- that is achieved even though we're using the
mainnet-trained dictionaries.

This change adds two dicts: one trained on a subset of blobs of mainnet
transaction and receipts data, and one on a subset of event blobs.

For this we have to enable using an experimental API in zstd that makes
it possible for our compressors and decompressors to reference
lazy-initialized dictionaries.