refactor: simplify migration by using block-based batching
Based on the previous state of the migration and the track record it had
it looks like iterating over all transactions and correctly accumulating
data for a single block is error prone.

We can simply fetch the data per-block instead thus simplifying the logic
we have in the migration code.