chore(github/docker): use `docker buildx imagetools` to tag with release tag
A `docker pull` / `docker push` pair does not properly keep the multi-arch
manifest list so that only the architecture the pull/push is run on gets tagged.

`docker buildx imagetools create` on the other hand communicates directly
with the registry and is capable of handling multi-arch manifest lists
properly. It also avoids having to download the image(s) themselves.