Feature: public function to compute Cairo 0 hinted class hash
Problem: we require a way to compute the "hinted class hash" used in the
Cairo 0 class hashing procedure for the Starknet OS. Pathfinder computes
this value but does not expose the computation.

Solution: extract the hash computation in a new public function
`compute_cairo_hinted_class_hash()`. Make the `json` module and the JSON
parsing function public as well to give callers access to the types
required to compute this hash.