feat(compiler): parse Sierra version from the class
And select which compiler version to use based on the Sierra version.

This makes it possible to compile classes without knowing the
StarknetVersion of the block they were declared in. We can finally
remove the special case in the executor code (that always used the
latest version of the compiler).