fix(pathfinder/sync/headers): temporarily disable header signature verification
Due to feeder gateway issues the length and commitment of
the state diff is not the same when returned by the
gateway as it is when we're generating the state diff.

Because of this the signature produced by the feeder
gateway may be invalid for affected blocks. This change
temporarily disables that check until the feeder gateway
is fixed.