fix(executor): don't duplicate messages from internal calls
When converting a blockifier `CallInfo` to a FUNCTION_INVOCATION we must
_not_ process all messages from inner calls recursively. Doing so results
in duplicating all messages from inner calls on the outer level, leading
to an inconsistency with the output from the feeder gateway.

Easy to reproduce with Sepolia testnet transaction
0x04faf2a913828ab965de32bf3004ce183e035e92ddbf9517772c5beaa4edb216 where
the message is emitted by an inner call and not the top-level contract.

Closes #2094