fix(storage/trie): fix off-by-one for no-history pruning mode
When inserting a new trie root (or contract state hash) we were first
pruning the old values (so that data that is out of the
number-of-blocks-kept window is removed). The pruning functions make
sure that there's at least one value at or before the first block of the
availability window available.

This was working correctly for non-zero number-of-blocks-kept values, but
was causing missed removals for the no-history case. When number-of-block-kept
is zero we were keeping the _previous_ value (root / state hash) in addition
to the new value we were being inserting.

This change fixes this by reordering insertion of the new node and
pruning. This way we correctly remove old values in the no-history case.

Closes #2110