chore: use rustls instead of native-tls
Previously we've been using native-tls which is a bit problematic on
Linux. Build time we're linking to the version of OpenSSL that is
available on the build host.

For example, Ubuntu 20.04 uses OpenSSL 1.1 so that the release binaries
we're building now are linked to that version of the library. That binary
won't work on Ubuntu 22.04 (and up) since those versions use OpenSSL 3.0.

This change switches `reqwest` to use `rustls` instead (with native
certs, meaning we're still using the same certificate store as OpenSSL).
This way we're no longer linking to OpenSSL and so our release binaries
will be more portable.