feat(storage/state_update): store re-declared Cairo 0 class information
This change adds a new `redeclared_classes` table storing block numbers
at which class re-declarations did happen.

When inserting a state update we take care of adding new rows here if
the state update contains a re-declaration -- and then when retrieving
the state update from storage we add re-declared classes to the set
of Cairo classes declared at the block.