fix(pathfinder/state/revert): always apply reverse state updates
When reverting Merkle tries to a previous state we have been checking
if we still have the tree root for the revert target. This seems like
a good idea, because if we _still_ have the trie(s) for the target
block then why bother applying reverse state updates from the current
state?

It turns out that's not so simple. First, trie nodes added in blocks that
are being deleted (right after the revert when pruning the block range
that has been reorged away) will be permanently leaked. There's no way
around that because we do not keep track of per-block trie node
_additions_, only _removals_.

Second -- and this is _much_ more problematic -- the way we were handling
trie removals data was just plain wrong. When reverting trie state to
block N we've just moved removals data for all blocks > N into our
target block. That removals data was effectively nodes that have been
deleted from the trie after block N. The problem with this is that
when we were then removing trie nodes based on the removals data of block
N later, we might be deleting nodes that were still reachable at that
block!

This change fixes this by removing the special case: when doing a revert
we are _always_ just applying a reverse diff to the _current_ state of
the trie. This makes sure that we'll compute and insert new state roots
for the target block and that the _removals_ data for that block will
indeed only contain nodes that are unreachable (either by having been
removed at a previous reorged-away block _or_ by the reverse update).

This change has been tested using the `feeder_gateway` tool triggering
thousands of multi-block reorgs while syncing Sepolia testnet block range
113k to 194k. Without this fix this end-to-end test was consistently
reproducing the "missing node" issue.

Closes: #2170